[XPU][MiniMax] Address review: kernels in src/sycl, benchmark vs HW peak
- Fold minimax_decode_topk.hpp into its single consumer MinimaxDecodeTopK.cpp
- Drop the torch eager providers; report achieved GB/s and % of device peak
- Make the bytes-moved model regime-aware (trivial path reads no scores)
- Register bench_minimax_decode_topk.py in the nightly benchmark suite
